tpch still doesn't run :(

+++ src/lib/types.hpp
@@ -134,7 +134,7 @@ using CommitID = uint32_t;
 using TransactionID = uint32_t;
 
 using StringLength = uint16_t;     // The length of column value strings must fit in this type.
-using ColumnNameLength = uint8_t;  // The length of column names must fit in this type.
+using ColumnNameLength = uint16_t;  // The length of column names must fit in this type.
 using AttributeVectorWidth = uint8_t;
 
 using PosList = pmr_vector<RowID>;
